@@ -13,43 +13,26 @@
     "record copy environment"
   ],
   "dependencies": {
-    "@datocms/cma-client-browser": "^1.0.1",
-    "@types/node": "^16.11.27",
-    "@types/react": "^17.0.44",
-    "@types/react-dom": "^17.0.15",
-    "datocms-plugin-sdk": "^0.5.2",
-    "datocms-react-ui": "^0.5.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.1",
-    "typescript": "^4.6.3"
+    "@datocms/cma-client-browser": "^5.2.5",
+    "@types/node": "^25.0.8",
+    "@types/react": "^19.2.8",
+    "@types/react-dom": "^19.2.3",
+    "datocms-plugin-sdk": "^2.1.1",
+    "datocms-react-ui": "^2.1.1",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
+    "typescript": "^5.9.3"
   },
   "scripts": {
-    "start": "cross-env BROWSER='none' PUBLIC_URL='/' react-scripts start",
-    "build": "cross-env PUBLIC_URL='.' react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "prepublishOnly": "npm run build"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
-    "cross-env": "^7.0.3"
+    "@vitejs/plugin-react": "^5.1.2",
+    "cross-env": "^10.1.0",
+    "vite": "^7.3.1"
   },
   "datoCmsPlugin": {
     "title": "Sidebar Record Copy Environment",
